Release 2.9.0
- 新模块 AntiBot 反假人
    - 仅限PAS模式。
    - 在玩家列表中标记出服务器中的假人。
- 重写 AntiVanish 反隐身
    - 仅限PAS模式。
    - 跟进最新的PAS服务器更改，反隐身失效问题得以解决。
- 新模块 Rotation 静默瞄准
    - 我们意识到旧的静默瞄准系统包含大量重复代码且相互冲突，易被反作弊误判。所以我们重写了静默瞄准系统。
    - 需要注意的是，如果关闭此模块，则每个模块的静默瞄准都不会工作。
- 新模块 Sprint 自动疾跑
    - tweakeroo所包含的自动疾跑并不合法。这个模块仅仅模拟按下疾跑键。
- 新修复 InvalidYaw 无效偏航角
    - Grim 反作弊有时会错误旋转你的头，过大的偏航角会导致你的游戏崩溃。
- 移除 NoFall 模块
- 移除 BlockDetector 模块
- 移除 Backtrack 模块
- 移除 ClickSimulator 系统
- 移除 SprintA 修复
    - 已经被合并到新的静默瞄准系统。
- 修复一个与ReplayMod崩溃问题。
- 一些细节修改。

+++ src/main/java/top/infsky/cheatdetector/impl/modules/common/AimAssist.java
@@ -47,7 +47,7 @@ public class AimAssist extends Module {
     }
 
     private void rotate(Entity target) {
-        Pair<Double, Double> rot;
+        Pair<Float, Float> rot;
 
         if (Advanced3Config.aimAssistInteract) {
             rot = new Pair<>(
@@ -58,12 +58,12 @@ public class AimAssist extends Module {
             if (Advanced3Config.aimAssistStopOnTarget && TRPlayer.CLIENT.crosshairPickEntity == target) return;
 
             rot = AimSimulator.getLegitAim(target, player,
-                    new Pair<>(Advanced3Config.aimAssistYawSpeed, Advanced3Config.aimAssistPitchSpeed),
-                    new Triplet<>(Advanced3Config.aimAssistOffsetX, Advanced3Config.aimAssistOffsetY, Advanced3Config.aimAssistOffsetZ),
+                    new Pair<>((float) Advanced3Config.aimAssistYawSpeed, (float) Advanced3Config.aimAssistPitchSpeed),
+                    new Triplet<>((float) Advanced3Config.aimAssistOffsetX, (float) Advanced3Config.aimAssistOffsetY, (float) Advanced3Config.aimAssistOffsetZ),
                     Advanced3Config.aimAssistNoise1,
-                    new Pair<>(Advanced3Config.aimAssistYawRandom, Advanced3Config.aimAssistPitchRandom),
+                    new Pair<>((float) Advanced3Config.aimAssistYawRandom, (float) Advanced3Config.aimAssistPitchRandom),
                     Advanced3Config.aimAssistNoise2,
-                    new Pair<>(Advanced3Config.aimAssistXZRandom, Advanced3Config.aimAssistYRandom)
+                    new Pair<>((float) Advanced3Config.aimAssistXZRandom, (float) Advanced3Config.aimAssistYRandom)
             );
         }
